Move filter in sub namespace
