+++ Macropus.ECS/Systems/IReactiveSystem.cs
@@ -1,4 +1,5 @@
 ﻿using Macropus.ECS.Entity;
+using Macropus.ECS.Systems.Filter;
 
 namespace Macropus.ECS.Systems;
