+++ Tests/ECS/BasicFunctionality/Systems/BrokenFilterTestComponentSystem.cs
@@ -1,5 +1,6 @@
 ﻿using Macropus.ECS.Entity;
 using Macropus.ECS.Systems;
+using Macropus.ECS.Systems.Filter;
 
 namespace Tests.ECS.BasicFunctionality.Systems;
