+++ Tests/ECS/BasicFunctionality/Systems/FilterTestComponentSystem.cs
@@ -1,6 +1,7 @@
 ﻿using Macropus.ECS.Entity;
 using Macropus.ECS.Systems;
 using Macropus.ECS.Systems.Extensions;
+using Macropus.ECS.Systems.Filter;
 using Tests.ECS.BasicFunctionality.Components;
 
 namespace Tests.ECS.BasicFunctionality.Systems;
