+++ Tests/ECS/BasicFunctionality/Systems/RemoveReadOnlyComponentSystem.cs
@@ -1,5 +1,6 @@
 ﻿using Macropus.ECS.Entity;
 using Macropus.ECS.Systems;
+using Macropus.ECS.Systems.Filter;
 using Tests.ECS.BasicFunctionality.Components;
 
 namespace Tests.ECS.BasicFunctionality.Systems;
